front-end/src/app/shared/buscador-transportista/buscador-transportista.component.ts
import { Component, OnInit, ViewChild, TemplateRef, ElementRef, Input } from '@angular/core';
import { NG_VALUE_ACCESSOR, FormBuilder } from '@angular/forms';
import { SearchComponent } from '../../core/shared/super/search.component';
import { Transportista } from '../data-models/transportista';
import { TransportistaService } from './transportista.service';
import { PopupService } from '../../core/services/popupService/popup.service';
import { Resources } from '../../../locale/artifacts/resources';

@Component({
  selector: 'yrd-buscador-transportista',
  templateUrl: './buscador-transportista.component.html',
  styleUrls: ['./buscador-transportista.component.css'],
  providers: [
    { provide: NG_VALUE_ACCESSOR, multi: true, useExisting: BuscadorTransportistaComponent }
  ]
})
export class BuscadorTransportistaComponent extends SearchComponent<Transportista, TransportistaService>
  implements OnInit {

  @Input() etiqueta = 'Transportista';
  @Input() cssClassEtiqueta = 'col-sm-3';
  @Input() cssClassControl = 'col-sm-4';
  @ViewChild('inputTransportista') inputElement: ElementRef;
  @ViewChild('selectCellTemplate') selectCellTemplate: TemplateRef<any>;

  constructor(transportistaService: TransportistaService,
              popupService: PopupService,
              private readonly fb: FormBuilder) {
    super(transportistaService, popupService);
  }

  ngOnInit() {
    this.baseElement = this.inputElement;

    this.advancedSearchForm = this.fb.group({
      cuit: '',
      descripcion: ''
    });

    this.columns = [
      {
        prop: 'selected',
        name: '',
        sortable: false,
        canAutoResize: false,
        draggable: false,
        resizable: false,
        cellTemplate: this.selectCellTemplate,
        width: 30,
      },
      {
        name: Resources.Labels.Cuit,
        prop: 'codigoFiscal'
      },
      {
        name: Resources.Labels.Descripcion,
        prop: 'descripcion'
      }
    ];

    this.subscribeToAdvancedSearchFormChanges('cuit', 'codigoFiscal');
    this.subscribeToAdvancedSearchFormChanges('descripcion', 'descripcion');
  }
}
